requirements.txt
requests>=2.32.3
urllib3>=2.2.2
python-dotenv>=1.0.1
psycopg[binary]>=3.2.1
Flask>=3.0.3

